internal/log: Ignore logger close error
Closing the logger is usually the last operation that will be run and at
this point, we don't have a good way to report errors, and it's also too
late to act on them. Thus ignore those errors.

Fixes lint:
```
Error return value of `l.ops.Close` is not checked (errcheck)
```