internal/exec/stages/disks/luks: Log previously ignored errors
Those errors are unlikely to happen and neither is fatal nor a bug as
the file is written in a temporary folder in the initramfs that is
removed before moving to the real root.

Thus log them and ignore them for as we don't want to start failing on
those if we did not before.

Fixes lint:
```
Error return value of `keyFile.Close` is not checked (errcheck)
Error return value of `os.Remove` is not checked (errcheck)
```