tests/filesystem: Join cleanup errors
Fixes lints:
```
Error return value of `out.Close` is not checked (errcheck)
Error return value of `os.RemoveAll` is not checked (errcheck)
Error return value of `writer.Flush` is not checked (errcheck)
```

It's fine to take into account close errors here as those are unlikely
and in this case those are in tests, thus not a potential subtle change
in behavior.

We also setup the defer'ed close earlier, as soon as the file is opened.