Support providing translating Butane configs via library
This change allows users to specify Butane configs [1] [2] directly
to their instances on startup instead of an Ignition config.

Many users already store their configs in Butane format and Butane
provides some "sugar" for making it easier to generate Ignition
specification configs. Some of this "sugar" is distro agnostic and
some of it is distro specific.

Implementing this means that users don't have to separately convert
their Butane YAML configs into Ignition JSON configs via the `butane`
CLI tool. This has the benefit of being easier for new users to
understand and also the benefit of not having to have intermediate
formats be either generated out of band at runtime OR stored in
version control. It has the drawback, though, of allowing users to
provide a configuration to their instance that is invalid and could
have been detected earlier.

After six years of experience in this space I believe the tradeoffs
mentioned above are worth making a change like this.

[1] https://coreos.github.io/butane/
[2] https://github.com/coreos/butane