internal: Upgrade to aws-sdk-go-v2
Update the code from aws-sdk-go to aws-sdk-go-v2, aws-sdk-go is out-of-support
since July 2025. I've tried to preserve the existing behavior as best I could:
- if we're running on EC2 we use ec2rolecreds first, with a fallback to anonymous
  credentials
- use anonymous credentials when outside EC2
- use IMDS for region hints

One subtle change appears to be that the S3 code does not tolerate leading
slashes in key names any more, so we need to strip it out explicitly.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>