url: Retry s3 fetch only when EC2 role fetch fails
This aligns the aws-sdk-go-v2 version with what we used to do for aws-sdk-go,
although the error matching is string-based now because the SDK changed its
error reporting.

When we're not on EC2 we directly fetch with anonymous credentials, so there is
no point retrying. On errors other than EC2 role fetch, such as when key is
invalid, it also doesn't make sense to retry.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>