Add generation and registration of the AK
Generate and persistent the AK in the TPM if the key registration is
request in the config. The public AK is then registered at the url
specified in the config.

The AK is generate only once and if the /var/tpm/ak.pub exists then it
means that this step has already been performed and it is skipped of the
next stages.

The registration of the AK required networking hence, during the
fetch-offline stage it signals that the networking is necessary.

The retries mechanism ensures that the registration is tried multiple
times for allowing the parallel network configuration in the fetch
phase.

Signed-off-by: Alice Frosi <afrosi@redhat.com>